BACKLOG-10018: fix test for emr310, move common logic for CommonHBaseConnectionTest to common source folder, separate for 1.0, 1.1 to  separate tests
